requirements.txt
hyperliquid-python-sdk
eth_account
duckdb>=1.1,<1.3
pytz
pandas
numpy
matplotlib
